scripts/requirements.txt
altgraph==0.17.3
distro==1.8.0
numpy==1.25.1
pandas==2.0.3
pefile==2023.2.7
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.2
six==1.16.0
tabula-py==2.7.0
tzdata==2023.3
